+++ guillotina/contrib/email_validation/utility.py
@@ -1,0 +1,133 @@
+from guillotina import app_settings
+from guillotina.component import get_utility
+from guillotina.contrib.email_validation.interfaces import IValidationSettings
+from guillotina.contrib.email_validation.utils import extract_validation_token
+from guillotina.contrib.email_validation.utils import generate_validation_token
+from guillotina.contrib.templates.interfaces import IJinjaUtility
+from guillotina.event import notify
+from guillotina.events import ValidationEvent
+from guillotina.interfaces import IMailer
+from guillotina.response import HTTPNotImplemented
+from guillotina.response import HTTPPreconditionFailed
+from guillotina.response import HTTPServiceUnavailable
+from guillotina.utils import get_registry
+from guillotina.utils import resolve_dotted_name
+from jsonschema import validate as jsonvalidate
+from jsonschema.exceptions import ValidationError
+
+import logging
+
+
+logger = logging.getLogger("guillotina.email_validation")
+
+
+class EmailValidationUtility:
+    def __init__(self, settings):
+        pass
+
+    async def initialize(self, app=None):
+        pass
+
+    async def finalize(self):
+        pass
+
+    async def start(
+        self,
+        as_user: str,
+        email: str,
+        from_user: str,
+        task_description: str,
+        task_id: str,
+        redirect_url=None,
+        context_description=None,
+        ttl=3660,
+        data=None,
+    ):
+        # from_user will be mostly anonymous
+        registry = await get_registry()
+        if registry is None:
+            logger.error("No registry")
+            raise HTTPServiceUnavailable()
+
+        config = registry.for_interface(IValidationSettings)
+        if config is None:
+            logger.error("No configuration on registry")
+            raise HTTPServiceUnavailable()
+
+        util = get_utility(IMailer)
+        if util is None:
+            logger.error("No mail service configured")
+            raise HTTPServiceUnavailable()
+
+        template_name = config["validation_template"]
+        site_url = config["site_url"]
+        validate_url = config["validation_url"]
+
+        render_util = get_utility(IJinjaUtility)
+        if render_util is None:
+            logger.error("Template render not enabled")
+            raise HTTPServiceUnavailable()
+
+        if task_id not in app_settings["auth_validation_tasks"]:
+            logger.error(f"Task {task_id} unavailable")
+            raise HTTPServiceUnavailable()
+
+        if data is None:
+            data = {}
+
+        data.update(
+            {"v_user": as_user, "v_querier": from_user, "v_task": task_id, "v_redirect_url": redirect_url}
+        )
+
+        token, last_date = await generate_validation_token(data, ttl=ttl)
+
+        link = f"{site_url}{validate_url}/{token}"
+        template = await render_util.render(
+            template_name,
+            context_description=context_description,
+            link=link,
+            last_date=last_date,
+            task=task_description,
+        )
+        await util.send(recipient=email, subject=task_description, html=template)
+
+    async def schema(self, token: str):
+        data = await extract_validation_token(token)
+
+        action = data.get("v_task")
+        if action in app_settings["auth_validation_tasks"]:
+            return app_settings["auth_validation_tasks"][action]["schema"]
+        else:
+            return None
+
+    async def finish(self, token: str, payload=None):
+        data = await extract_validation_token(token)
+
+        action = data.get("v_task")
+        if action in app_settings["auth_validation_tasks"]:
+            if "schema" in app_settings["auth_validation_tasks"][action]:
+                schema = app_settings["auth_validation_tasks"][action]["schema"]
+
+                try:
+                    jsonvalidate(instance=payload, schema=schema)
+                except ValidationError as e:
+                    raise HTTPPreconditionFailed(
+                        content={
+                            "reason": "json schema validation error",
+                            "message": e.message,
+                            "validator": e.validator,
+                            "validator_value": e.validator_value,
+                            "path": [i for i in e.path],
+                            "schema_path": [i for i in e.schema_path],
+                            "schema": schema,
+                        }
+                    )
+
+            task = resolve_dotted_name(app_settings["auth_validation_tasks"][action]["executor"])
+
+            result = await task.run(data, payload)
+        else:
+            logger.error(f"Invalid task {action}")
+            raise HTTPNotImplemented()
+        await notify(ValidationEvent(data))
+        return result