+++ guillotina/contrib/templates/content.py
@@ -1,0 +1,14 @@
+from guillotina import configure
+from guillotina.content import Resource
+from guillotina.contrib.templates.interfaces import IJinjaTemplate
+
+
+@configure.contenttype(
+    type_name="JinjaTemplate",
+    schema=IJinjaTemplate,
+    add_permission="guillotina.AddJinjaTemplate",
+    behaviors=["guillotina.behaviors.interfaces.IDublinCore"],
+    allowed_types=[],
+)
+class JinjaTemplate(Resource):
+    pass