+++ stubs/jsonschema/__init__.py
@@ -1,2 +1,7 @@
 class ValidationError(Exception):
     ...
+
+
+def validate(instance, schema):
+    ...
+